requirements.txt
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
Flask>=3.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.66.0
